Angelos' Fuji camera is an 8070i device and it only accepts 12 byte CDBs. I'm going to guess that ATAPI and QIC157 devices also want 12-byte CDBs.
So, pad CDBs to be 12 bytes for ATAPI, QIC157, 8070i (as well as UFI devices)

Also, remove an overly paranoid check in BBB state machine which
actually violates the spec.  This is half of getting Angelos' Olympus
camera to work. The other half involves a semi-violent rewrite of the
sd scsi driver to make it more tolerant.