Hardware clock support, via adb. This reduces the dependancy on openfirmware somewhat. Since Openfirmware is not actually used for anything, it is no longer necessary to configure it. so several lines of useless dmesg are now gone.
This does not have clock setting code enabled yet.
All of the code is in place, but needs further testing before
it is trusted.

I wish Apple would store UTC not localtime in the hardware clock,
besides the fact that the clock base is 1904.

Need to keep the clock sane for dual boot machines.