Tweak softdep_change_linkcnt() so we can specify whether we're willing to be co-opted by softdep or not, and use this new interface to inform softdep in ufs_inactive() that the inode mode has changed.
We don't want to be co-opted there as that might lead to undesired
circular dependencies such as a vput() depending on another vput() to
complete, or a process that is trying to free up a vnode being blocked
trying to acquire a new vnode.

Okay tedu@ deraadt@, thanks to all those who tested.