back out 1.97. it was causing a freeze on boot with an older card. problem found and cause narrowed down by jrrs.
it looks like we're trying to submit a command during attach to be
completed by polling. however it must be generating an interrupt for the
command, which completes out of the interrupt handler behind the scenes
of our polling. we just seem to sit there forever as a consequence.

this is a bandaid over a real fix, but it'll do until i can talk marco into
fixing it properly.