MFC: Fix by dlg@
the megaraid i4 and my dell perc 3/dc share the same vendor and product
ids, but the i4 is considered broken while my perc is definately working
fine. so instead of marking all the devices with this vendor and product
id as broken, we use the pci subsys ids to determine if the device needs
a quirk.

ok deraadt@ marco@ dlg@