define the command structs used by the ports.
this stuff isnt endearing me to ata at all, and apparently ahci is a good
ata controller. the command path looks way more complicated than it needs
to be (or maybe i am used to really good scsi controllers, i dont know). a
lot of the terminology seems annoyingly different just for the sake of
being different. i mean, what is the advantage of calling a scatter gather
lement a "physical region descriptor table"? noone else calls it that.
there's a lot of weird stuff in here.

blah.