Allows any input/output stream to underrun/overrun without disturbing other streams. Beside making aucat usable with slow apps (eg. cdio), this change will ease turning aucat into an audio server later
if one of the input buffers underruns, then silence is generated in
place of the missing samples. Later, as many samples are dropped in
order to maintain the stream in sync with the rest

if one of the output buffers overruns then newer samples are
discarded. Later silence is generated in order to maintain the
stream in sync with the rest.

ok jakemsr