Virtually re-implement origination of Intra-Area-Prefix LSAs. We now comply with RFC5340 and move prefixes between the two ref-types of Intra-Area-Prefix LSAs (router and network).
In case an LSA ends up having all of its prefixes moved to another
one, the LSA is flushed from the LSDB by re-originating with MAX_AGE.
(This does not yet work reliably. We end up having LSAs with zero
prefixes in the LSDB in some cases. Should not affect SPF calculation,
so not much harm done, but will need to be fixed.)

This commit just adds the necessary code to do origination,
but it is not being called yet. This will be done in a separate
commit.

"Looks good." claudio@