Send valid ERROR packets to prematurely terminate a transfer.
According to RFC1350 there should be always be a terminating NUL.
Andre Gillibert on bugs@ pointed out that the misformed packets
caused Gentoo Linux's tftpd (atftpd) to crash trying to transfer
pxeboot to an OpenBSD machine.

This is slightly different diff than the one proposed by Andre, but
he confirms it also fixes the problem.

"looks like a better diff to me" deraadt@.