A couple of  fixes (based mostly on stuff from upstream) to stop breaking vblanks over suspend.
Firstly, when we turn off the irq, wait up anyone waiting on vblanks,
and prepare for the fact that interrupts are going off.

Secondly, only reduce the refcount for vblanks over modeset if they were
actually turned on.

Fixes gl apps running while you suspend an x40 (as in they still work
when it comes back).