when requesting MX entries, the result can be appear in random orders.
the logic for inserting them in a lka session when acting as backup MX did
not take account for one specific case that could lead to an early exit
without smtpd getting a chance to detect the entry corrsponds to itself.

in such case, a backup MX woud try to connect to itself and bounce in the
loop detection code ... or it would sometimes work.