Extend the keyboard communication routines to be able to work in polling mode; use this to read the DIP switches from the keyboard at attach time.
Change the state machine to allow for a `DIP switch prefix' scan code to be
received while we are not attempting to read the DIP switches, for the
`international' key (not found in regular us layouts, documented is the
`GERlessthan' key in sgi's keyboard(7) manual page) will return the
aforementioned scancode, instead of the one documented in the manual.

Thanks to sebastia@ for lending me his german layout keyboard.