gcc4: emit warning when ignoring alignment constraints
Currently, GCC 4.2 silently ignores the "aligned" attribute for
objects allocated on the stack if the specified minimum alignment
exceeds the platform's natural stack alignment.  This has bitten us in
the past, so we shouldn't allow this to continue.

Fixing the "ignores" problem seems hard, so this commit settles for
tackling the "silently" problem instead.

ok miod, and possibly guenther and deraadt