When running on a sun4d or sun4m system, have the bootblocks ask for one more megabyte of virtual memory beyond the kernel image; this is safe to do as you can't find any such system with less than 8MB of contiguous memory.
While nothing in the kernel has a need for that extra memory to be available
to the kernel for its bootstrap needs, forthcoming changes (post-5.7) will
need it, and these kernels won't work when booted from boot loaders prior to
this change. By shipping 5.7 with updated boot blocks, we'll make upgrading
to 5.7-CURRENT easier for sun4m owners, when the time comes.

Crank boot blocks version.

ok deraadt@