Convert membar_* calls into virtio_membar_sync where it matters
membar_* functions are defined only as compiler barriers on !MP
kernels, while we're trying to be conservative in our use of the
barriers.  Barriers are placed only where loads and stores might
get reordered and it matters at the same time.  Shared info page
operations are using atomic instructions on Linux, so they get
barriers as well.