telnet should not verify if hostname is a fully qualified domain
telnet would perform a gethostbyname on close/quit commands
when the machine hostname is not a fully qualified domain.
This behavior added on 12-March-98 breaks the current pledge
that "dns" operations are not needed after establishing
a connection leading to a coredump from telnet when
quiting or closing an existing connection.

OK chris@