backport r1.75: Make sure that amap slot calculation does not overflow
This prevents too small amaps from being allocated by
forcing the allocation of a large number of slots.

Based on an analysis from Jesse Hertz and Tim Newsham.

This is a different patch for 5.9 that addresses the same
issue as r1.75 of uvm/uvm_amap.c. It also makes sure that
vmm(4) cannot make such large amap allocation requests.