When pmap_page_remove() is called by UVM, a physical page is to be removed from pmaps it currently is in.  To check if a virtual address pointing to that physical page has been mapped, the code uses the l2pte_valid() function.  Unfortunately there is a difference between being valid and the PTE being zero.  If a page is mapped but has never been accessed, it will be non-zero but invalid.
In that case the PTE for that virtual address will not be zeroed
and the virtual address will be removed from the vm page struct.

The next time someone tries to map a page to that virtual address,
other pmap code will consider the virtual address to be already
mapped, even though that assumption is completely wrong.

To make sure this does not happen, check the PTE for zero.  This way
the PTE will be zeroed correctly.  The check for zero is how other
ARM pmap code also handles this issue.

ok kettenis@ tom@