Enable hardware VLAN tagging
Turns out that hardware VLAN tagging is required to use VLANs at all
(while QinQ VLANs are not filtered regardless...). On top of that a
PowerShell command needs to be executed to configure Trunk mode on
virtual interfaces.  An example of such command is:

    Set-VMNetworkAdapterVlan -VMName OpenBSD -Trunk \
        -AllowedVlanIdList 10-100 -NativeVlanId 5

This will enable trunking mode on all hvn(4) interfaces.  An option
-VMNetworkAdapterName can be passed to turn trunking on one interface,
and although by default all networking interfaces have the same name
"Network Adapter", a "Rename-VMNetworkAdapter" command can be used
to rename them one by one.