Stop complaining about changed kernel options when called for an empty directory. With the new kernel obj mechanism config always prints
	Kernel options have changed -- you must run "make clean"

on the first run. The message is suppressed when the compile directory
didn't exist before, but now the directory always exists when config is
run, so we have to use something else to detect if this is a "fresh" run
or an update of an existing directory. Use the options file instead,
which is used to track the kernel options of the last config run. If the
file doesn't exist, it is safe to assume, that this is a "fresh"
directory.

ok tb deraadt