serialise posts to the txstart register.
without this it is possible for re_txeof to let the chip look at
and complete tx descriptors that re_start has written to the ring
but not accounted for in the software producer index. in this
situation no future interrupts may be generated for packets on the
ring, and tx will stall.

the serialisation is implemented by only doing the writes from tasks
via ifq_serialize.

this is part of a bigger diff that fixes lockups reported by haesbaert.