Regress test for sparse files & mmap with different FSs
This test creates a file with a hole in the middle and then checks that
when reading or mmaping the whole file, the data is correct.

This catches the bug introduced by the
"Implement VFS read clustering for MSDOSFS" commit that hat to be
reverted.

For now, we run the test with FAT16, FAT32, FFS

With much advice from bluhm@