Scale the missed beacon counter threshold to the AP's beacon interval.
This should make fading APs time out consistently regardless of what the
beacon interval is set to (range is 1 to 2^16 TU, though in practice 100 TU
seems to be a common value).

Print the beacon interval and missed beacon counter threshold to dmesg
if the DEBUG flag was set on the wireless interface with ifconfig(8).
This should help with diagnosing any issues that pop up.

Requested and diff eye-balled by kettenis@
help & ok tb@ phessler@