Two background scan fixes for iwn(4) (4965 devices only):
 1) Use only one antenna while scanning on 5GHz. Fixes very low RSSI values.

 2) During a background scan while associated to a 5GHz AP, send probe
    requests on 2GHz channels with an OFDM rate (6Mbps) because the buggy
    firmware does not like sending with a CCK rate (1Mbps) in this state.
    CCK rates are not valid for 5GHz, which could explain this firmware bug.

Taken from FreeBSD r222679.

jca@ tested and confirmed the first problem but lacks a 5GHz access point to
associate to and test the second one with.
I am going to *blindly trust* FreeBSD on the second one! It is not a big deal
if it doesn't actually apply to our firmware version as the change itself is
rather inconsequential in practice. And 4965 hardware is quite rare nowadays.

ok jca@