Initial version of Crypto Simplified Interface (CSI).
This is a code base that intends on providing a simplified interface for
mid-level cryptographic operations. In due course various applications and
libraries will be able to benefit from a clean and robust API, rather than
using libcrypto or other similar APIs directly.

Discussed at length with deraadt@, djm@, markus@, beck@ and others.