vmd(8): avoid a divide by zero when the user specified a low baud rate
Low baud rates would result in a 0 "rate limiter pause count" in the
serial port output code. This pause counter is used to implement a
delay in what otherwise is an instantaneous serial port output path, and
is needed by some guest OS kernels. This fix only enables the rate
limiter pause if the count is > 0 ("pause after 0 characters" makes no
sense anyway).

Note that this will result in skipping the limiter on unusually low
baud rates, but since nobody is using a low baud rate console in vmd,
I don't think this is going to be a problem.

ok pd