EVFILT_TIMER: Remove extra tick from tvtohz(9) on timeout reload.
tvtohz(9) adds an extra tick to account for the present tick, but this
tick needs to be removed when the timeout is reloaded thereafter.  We
already do this for periodic setitimer(2) timeouts.

Prompted by Paul Herman's writeup on clock aliasing for DragonflyBSD:
https://frenchfries.net/paul/dfly/nanosleep.html

Also fixed in FreeBSD r238424.

Style tweaks from visa.

ok visa@, guenther@