Correctly disable pvclock(4) on old hardware that lack a stable clock
I falsely assumed that the KVM_FEATURE_CLOCKSOURCE_STABLE_BIT
indicates that the actual clock values are stable, but it turned out
that this isn't always the case.  To detect if the clock value is
stable, we now read it once in pvclock_attach() and check for the
PVCLOCK_FLAG_TSC_STABLE flag.  This needs further investigation.

Reported and fix tested by johnw.mail at gmail.com

OK chris@ phessler@