Add handshake message type checking and special case certificate requests.
Check that the handshake message type received matches that required by the
state machine.

However, thanks to poor state design in the TLSv1.3 RFC, there is no way to
know if you're going to receive a certificate request message or not, hence
we have to special case it and teach the receive handler how to handle this
situation.

Discussed at length with beck@ and tb@ during the hackathon.

ok tb@