sync with kristaps, commit c5350b9055c3921001778275bb808a5c811c21c1
Protocol-level support for -o.  The actual sharing itself is more or less a
copy of the existing -g code.  I've refactored a lot of the ids.c code to be
generic and only accept a "isgid", as the rest is the same.  So there's some
churn as I make all references to "group" be generic.