When upgrading only prompt for disks that contain a root partition.
With this the question "Which disk is the root disk" has the correct
default in case of full disk encryption. Before it would list sd0 sd1
with sd0 being the backing store and sd1 the unlocked disk.

This is a step towards fully unattended upgrades running completely on
defaults (i.e. no need for an autoinstall(8) answer file).

In the past there have been concernes that we should not touch any odd
disk without the operator telling us as there might be bugs lurking.
It is early enough in the cycle to find out for sure.

input halex, kn
"just commit it" deraadt