Add more amdgpu device ids.
The marketing names in libdrm's amdgpu.ids are based on device id and
revision id.  A single device id often maps to multiple marketing names
but the same codename so stick to codenames here.  The marketing names
in amdgpu.ids also do not cover all the device ids amdgpu matches on.

Use the same codenames as amdgpu.
Polaris 20 and Polaris 30 as Polaris 10.
Polaris 21 as Polaris 11.
Polaris 22 as Vega M.