sprinkle some KERNEL_ASSERT_LOCKED()
operations that tweak the kq_head and kq_count need to be serialised
against the kevent syscall which also fumbles with the list and
count too.

these asserts would have made it extremely obvious where the tun(4)
bug was. for half the time of the bug report about it we werent
even sure it was tun(4)

discussed with mpi@ jmatthew@