handle ERSPAN type I
type I and II share their GRE protocol, but you tell them apart by
checking if a sequence number is used. type I does not use a sequence
number and contains a bare ethernet frame. type II contains a
sequence number and a shim header that is already handled by the
code.

tested with a Dell S5048F-ON and an encapsulated remote port mirror setup.