Fix merge(1) of files which contain lines with leading dots.
A check which tried to detect "dot on a line of its own" in the
internal ed script was treating any line with a leading dot as such.
Changes made to groups of such lines therefore ended up being
truncated in the merged result.

"For a random user, there's no shame in being unable to use ed."
	-- Michael W Lucas

ok millert@ deraadt@