gettimeofday, settimeofday(2): limit timezone support
For gettimeofday(2), always copy out an empty timezone struct.  For
settimeofday(2), still copyin(9) the struct but ignore the contents.

In gettimeofday(2)'s case we have not changed the original BSD semantics:
the kernel only tracks UTC time without an offset for DST, so a zeroed
timezone struct is the correct thing to return to the caller.

Future work could move these out into libc as stubs for clock_gettime and
clock_settime(2).  But, definitely a "later" thing, given that we are in
beta.

Update the manpage to de-emphasize the timezone parameters for these
syscalls.

Discussed with tedu@, deraadt@, millert@, kettenis@, yasuoka@, jca@, and
guenther@.  Tested by job@.  Ports input from jca@ and sthen@.  Manpage
input from jca@.

ok jca@ deraadt@