let lacp mode trunk cope with interfaces that don't report if_baudrate.
jmatthew@ and i thought i'd broken lacp in trunk(4) when adding
aggr(4), but we couldnt see how. turns out we were trying lacp mode
trunk only on new drivers (mcx and ixl) that didn't set if_baudrate,
which tickled an edge case in trunk that prevented it from selected
those interfaces for an aggregation. ixl and mcx have since been
fixed, but there's no reason for trunk to be this picky when we
make this ommission again.

ok deraadt@