scsi_delay(): sleep without lbolt
If we want to sleep for a multiple of seconds we can do that without
involving lbolt.

This may cause some paths to sleep longer than they have on average,
as sleeping on lbolt wakes you up within one second, not after one
second.  If this is a problem we will need to shorten the intervals
given to scsi_delay().

With insight from deraadt@.

ok krw@