Add iwx(4), a new driver for Intel AX200 wifi devices.
The iwx(4) driver is based on code from iwm(4) and Linux iwlwifi.
We are using a separate driver for this hardware family to avoid introducing
even more complexity to iwm(4). Firmware loading and the Tx machinery in
particular work differently compared to iwm(4) devices.

Device firmware can be installed with fw_update(8).
These devices support Tx rate scaling and Tx aggregation in firmware but we
leave this disabled for now and adapt the Tx rate with MiRa/AMRR instead.

Active scans (scans with probe requests) and background scans don't work yet.
Otherwise this driver has the same feature set as iwm(4).