Fix MiRA's sub-frame error rate computation.
When describing the equation for computing a sub-frame error rate (SFER) the
MiRA paper uses "hardware retries" to refer to the number of times a frame,
which eventually failed entirely, was retransmitted (a situation our drivers
would call "txfail"), and "nBad" to refer to frames which were "received
with errors" (and this is what our drivers would call "retries").

Because of my misunderstanding of the wording used in the paper our
SFER equation ended up with the "retries" and "nBad" variables swapped.
Swapping them back produces more meaningful results with reasonable
frame loss percentages based on the counters passed in by drivers.

ok tb@