Update MiRA probing interval of probed rates which are worse than the current best rate, not worse than the rate currently being probed.
This seems to be a more accurate interpretation of the MiRA paper.

The paper says the interval for a rate needs to be updated if the rate's
goodput is worse than that of the "current transmission rate" (see the
"Adaptive probing interval" section). Our implementation interpreted
"current transmission rate" as "rate being probed right now" and adjusted
the interval of the previously probed rate. However, the context of this
section of the paper suggests that "current transmissions rate" intends to
refer to the currently selected best rate for our non-probing transmissions.

testing and ok tb@ jmatthew@