correct mutex_lock_interruptible()
Linux kernel code often passes errors around as negative numbers cast to
pointers.  As rw_enter() returns a errno on failure
mutex_lock_interruptible() negated the return value.  But this did not
account for ERESTART being -1 which would return 1 to the caller.

sthen@ periodically hit a uvm_fault() in i915_request_create() which
was caused by attempting to use 1 as a pointer.

ok kettenis@