Fix unconditional write to v1.scan_priority in iwx(4) scan command.
The data structure for the scan command contains a union which represents
various versions of the scan command. The driver wrote to the scan_priority
field in version 1 data regardless of the scan command version actually
expected by firmware.

Perform that write only for scan command version 1, as intended.
Testing suggests that this prevents occasional firmware errors during scans.