Add mvpp(4), a driver for the Marvell Packet Processor v2, as used on the Armada 7K and 8K SoCs.  Additionally, it seems to be used on the upcoming CN9K SoCs.  This only supports version v2.2, since the older v2.1 is only used on some old ARMv7 we have no support for, and where there is no hardware for us to support.
A huge part of this driver is configuring the TCAM/SRAM-based "parser".
Thus most of the code, especially the parser handling, was ported from
EDK2 with a sed script doing plenty of automatic renaming and un-camel-
casing.

The controller supports multiple TX/RX queues/vectors and RSS, so this
would be a nice platform for testing network processing improvement on
arm64.  For now though we only configure and use a single queue.

The driver is still unfinished, but it's time to continue the work in
tree.  At least we can already send and receive packets, good enough
for dhclient to work.

Since the MACCHIATObin uses Clause 45 10G PHYs and SFPs, link state
handling is still a bit bogus, as we have no support for that.  The
1G RJ45 port works fine though.

ok dlg@