Add size to free(9) calls
import_identities() calls import_identity() which allocates a buffer and
potentially frees it itself;  if not, import_identities() uses it and frees
it afterwards.

Instead of crunching down the buffer size twice, make import_identity()
calculate and pass it back, similar to how pfkeyv2.c:pfkeyv2_get() does it.

Tested and OK tobhe