setitimer(2): ITIMER_REAL: don't call timeout_del(9) before timeout_add(9)
If we're replacing the current ITIMER_REAL timer with a new one we
don't need to call timeout_del(9) before calling timeout_add(9).
timeout_add(9) does the work of timeout_del(9) implicitly if the
timeout in question is already pending.

This saves us an extra trip through the timeout_mutex.