Fix rpki-client regress
The fix for the misuse of EVP_PKEY_cmp() (rpki-client/cert.c -r1.16)
came with an API change. ta_parse() will now throw an error if fed a
NULL pubkey. This in turn broke a regress test.

Fix this by parsing the pubkey out of the appropriate TAL to let ta_parse()
verify that it matches the pubkey in the first level certificate.

Discussed with tobhe, benno and claudio