Add "netstat -R" example
It's handy and otherwise easily missed when reading up on routing domains
and tables;  wording taken from netstat(1) as is.

Not listing pgrep(1)'s `-T' because examples don't have to be exhaustive and
ps(1) is already demonstrated;  same for top(1) users which more likely come
across its `t' and `T' in the help page anyway (I guess).

OK jmc remi