Fix segfault in pstat -v
Broken in r1.122 when the vnode list at the mount point was converted to
a TAILQ to make softdeps happy.  There was a for loop that looked a lot
like a LIST_FOREACH that was converted to a TAILQ_FOREACH. Unfortunately,
the loop is a bit more intricate. Revert to the original loop logic, but
now with TAILQ.

"looks correct" millert, "looks good" deraadt