Fix some issues with referencing named ACPI nodes from Packages. These references need to be resolved at runtime rather than when they're parsed such that they pick up the right values for those nodes which can be changed when for example _INI methods run.
The current approach is to replace these reference with a string that
names the node in question.  The problem with that is that packages
can also contain normal strings.  Which means that we need special
code that depends on the context where the package is used.

This diff takes a different approach by retaining a reference when
parsing.  Code that uses a package will need to resolve this reference
but there is no ambiguiety anymore.

ok patrick@