this is errata 6.8/011_nd6.patch.sig
date: 2021/01/06 08:10:15;  author: florian;  state: Exp;  lines: +3 -1;  commitid: r6m6UUdnnHvJu4Or;
Mark a layer 2 address as expired by setting the length to 0 when the
associated IPv6 NDP entry is invalidated. Otherwise we end up with an
INCOMPLETE entry that can't be updated to STALE and REACHABLE by
neighbor advertisements and thus interrupting communication.
This is the same as arpinvalidate() for IPv4.
Guidance bluhm & claudio, fix proposed by claudio
OK claudio