create /dev/ drm nodes with the same names as linux
This was proposed by Emil Velikov to simplify libdrm and will remove the
need for some patches in ports.

/dev/drm0 -> /dev/dri/card0
/dev/drmR128 -> /dev/dri/renderD128

The previous names will remain for a period of time and will later be
removed.  Major and minor numbers remain the same.

libdrm will not be changed to use the new names until known privsep
and sandbox use has been updated to allow the new names.

ok deraadt@