Since iwi(4) doesn't call into net80211_newstate() the interface link state must be updated by the driver in order to get packets to flow.
In case of WPA the link state was updated as a side-effect of a successful
WPA handshake. This commit fixes the WEP and plaintext cases.

Similar fix as recently committed to ipw(4).

Additionally, check for errors from iwi_auth_and_assoc() and keep scanning
if this function fails.

Problem confirmed and fix tested by matthieu@
ok deraadt@