Bugfix: when X509_NAME_dup(3) failed, X509_NAME_set(3) indicated success even though it did not actually set the name. Instead, indicate failure in this case.
This commit sneaks in a small, unrelated change in behaviour.
If the first argument of X509_NAME_set(3) was NULL, the function
used to return failure.  Now it crashes the program by accessing
the NULL pointer, for compatibility with the same change in OpenSSL.

This merges the following two commits from the OpenSSL-1.1.1 branch,
which is still available under a free license:
1. 180794c5  Rich Salz        Sep  3 11:33:34 2017 -0400
2. c1c1783d  Richard Levitte  May 17 09:53:14 2018 +0200

OK tb@