tr(1): plug leak in genclass()
If we have already generated a given character class we don't need to
do it again.  We can also return some of the memory we allocated for
the class.  NCHARS is an upper bound, most character classes are
small.

This fixes a small leak in genclass().

While here, switch to an ANSI function definition.

Thread: https://marc.info/?l=openbsd-tech&m=163571942030440&w=2

ok millert@