uniq(1): freopen(3) stdin, stdout
uniq(1) defaults to stdin and stdout, has exactly one input and one
output, and permits the user to replace the default input and/or the
default output with a given file.

freopen(3) is the right idiom in this context.  Get rid of file() and
the extra FILE pointers.  Switch from fprintf(3) to printf(3).

While here, const the "str" argument to show().

Thread: https://marc.info/?l=openbsd-tech&m=164078603312936&w=2

ok millert@