add a basic printer for EAPOL packets.
EAPOL turns out to be a little container for a bunch of other types
of packets, including EAP for use with vanilla 802.1X, a kind of
capability announcement thing, and MACsec Key Agreement.

it's not worth adding a separately file for such a small chunk of
functionality, and it seems specific to ethernet. parsers/printers
for the sub protocols can come later if needed.

ok deraadt@ visa@