Call uvm_pglistfree(9) instead of uvm_pmr_freepageq().
There is no functionnal change as the former is just a wrapper around the
latter.  However upper layer of UVM do not need to mess with the internals
of the page allocator.

This will also help when a page cache will be introduced to reduce contention
on the global mutex serializing acess to pmemrange's data.

ok kettenis@, kn@, tb@