Don't print "You may wish to update your font path" at pkg_add time for paths which are under /usr/local/share/fonts.
In most cases only fontconfig is relevant and this automatically
searches subdirectories, so user changes are not often needed any more.

The case where they might be needed is for server-side bit-map fonts
(managed via 'xset fp' or 'fontpath' in /etc/X11/xorg.conf) but this is
not common any more (most fonts in ports are not bit-map anyway) and
this gets rid of some cases of pkg_add messages which the user does
not need to act upon.

ok landry@ matthieu@ and I think there were some others