Delete the statement that the default stylesheet only used CSS1 because that has no longer been true for some time now.
I would certainly like to adhere to a coherent standard and state
which one that is.  Unfortunately, the W3C deliberately smashed
the CSS standard into pieces such that a coherent standard no
longer exists and such that statements about standard conformance
have become next to meaningless.  Consequently, i now remain
reluctantly silent regarding CSS standard(s) conformance.

Going back to CSS2.1, published in 2011, which was the last CSS
standard in the proper sense of the word, is not an option because
it has gaping holes in functionality and is no longer adequate for
use on today's WWW.