Replace selwakeup() with KNOTE() in socket event activation
Let's try this again now that the kernel locking issue in nfsrv_rcv()
has been fixed.

The previous attempt of the conversion triggered hangs on NFS servers.
This was probably caused by the removal of the kernel-locked section
just prior to the socket upcall. The section had masked a locking error
in NFS code.