fix dwim for reversed patches
patch(1) fails to recognize the reversal application of a patch that
cerates a file.  since an empty context always matches, the idea is to
run the dwim ("do what I mean") code also when locate_hunk succeeds but
the patch would create a file and the match is on the first line.
fixes the (disabled) test t3.

ok stsp@