sed(1) first appeared outside of Bell Labs in PWB/UNIX 1.0
it did not start in the PWB group

"The talk said that tools like grep and sed came from PWB,
but that's not true. They were original"

"The flow from PWB back to the main research line was a trickle at best.
We had bad NIH in 1127."

Rob Pike
https://minnie.tuhs.org/pipermail/tuhs/2020-February/020329.html

The 4.4BSD version was written by Diomidis Spinellis
credited in csrg/admin/admin/contrib

"In 1992, as a bored PhD student, I reimplemented sed(1) and contributed
it the unencumbered BSD version that was then being put together"
https://www.spinellis.gr/blog/FreeBSD.html

with and ok schwarze@