Encode an ASN.1 INTEGER with NULL data to value of zero.
When an ASN1_INTEGER is created it has NULL data until a value is set -
previously, an ASN1_INTEGER in this state encoded to an ASN.1 INTEGER with
a value of 0, rather than being treated as an error. While code should
really set values, the historical behaviour has not required this.

Found the hard way by sthen@ with acme-client.

ok tb@