Remove unused AUDIO_PROP_{MMAP,INDEPENDENT}
AUDIO_PROP_FULLDUPLEX is the only audio(9) in use, the other two died with
	commit 1cf2860827c8ca659d8097d8da94a5ae5b888c53
	Author: ratchov <ratchov@openbsd.org>
	Date:   Thu Jun 25 06:43:45 2015 +0000

	Reimplement the audio driver in a simpler way, removing unused/unusable
	functionality. Same API and ABI except for the removed bits and no
	behaviour change for programs using libsndio. With help from armani@
	and mpi@, thanks.

but remained defined and set in drivers.

and the following merely moved them when they were dead code already:

	commit 9215aa3dfad387bca877a805534df6dcfe8722eb
	Author: ratchov <ratchov@openbsd.org>
	Date:   Wed Aug 31 07:22:43 2016 +0000

	    Delete unused ioctls and associated macros. Move macros that are still
	    used internally by low-level drivers from sys/audioio.h to
	    dev/audio_if.h instead of deleting them.

None of this is used in base or ports;  codesearch.debian.net only shows
AUDIO_PROP_{CAPTURE,PLAYBACK} in firefox-esr, mozjs and cubeb.

ratchov points out that audio_if.h and audioio.h are private interfaces
and the codesearch shows SunOS and NetBSD bits (#ifdef'd out on OpenBSD).

OK ratchov