Limit IFS scope to effect parsing code only
Set to split `hw.disknames' strings is neat, but functionn-wide means the
bsort() invocation honours it, i.e. it'll output newlines not spaces.

Break the one-liner into the same multi-line idom occuring elsewhere and
contain IFS in the subshell that needs it.

This was never visible due how the output of was consumed by the shell.

The only visual, but otherwise effectless bug due to this appeared with the
new disk encryption question double-quoting get_dkdevs() output, i.e.
multiple disks printed across multiple lines;  this yields one, as inteded.

OK afresh1