Support upstreamed AP806/CP110 bindings in mvtemp(4).  Apparently older EDK2 versions shipped bindings compatible with the Armada 380, but when the ARM64 SoCs were upstreamed from the vendor fork to Linux, those apparently were adjusted.
With this diff mvtemp(4) now tries and use the parent as syscon regmap in case
the legacy binding cannot be used.  Furthermore this adds some initialization
as apparently the new EDK2 does less of that work during bootup.

Noticed by mbuhl@ who's running an updated EDK2 on the MACCHIATObin.
ok kettenis@