Treat crypto disk like the root disk, both are boot disks
Chosing an existing OpenBSD partition on GPT during disk setup requires an
existing EFI Sys partition (the installer does not create it in thi case).

With 'Encrypt the root disk?' answered postively, the crypto disk instead of
the root disk becomes the boot disk.

Extend the EFI Sys sanity check to both crypto and root disk, really asking
"is this a boot disk?".