Implement support for pointer authentication (PAC) in userland.  With PAC it is possible to "sign" pointers with a hidden key.  The signature is placed in unused bits of the pointer and can be checked later.  This can be used to provide "tail CFI" that is similar to what retguard provides.
Debuggers need to be aware of the fact that pointers can be signed.  For
this purpose a new PT_PACMASK ptrace(2) request is introduced that returns
as mask that indicates the bits used for the signature.  Separate masks
are provided for code and data pointers even though the masks are identical
in the current implementation.  These masks are also written into a special
note section in the core dump.

ok patrick@