explicitly disable Energy-Efficient Ethernet (EEE) on Marvell E151x
The Energy Detect feature is part of EEE and it defaults to off on both
hardware and software reset to enter "normal 10/100/1000 Mbps operation."
This change makes it explicit, as for the other PHYs in that family.

Note that, if EEE was actually enabled, an errata applies which requires
a certain sequence of magic register writes.

ok stsp@ kettenis@