unbreak parsing of IPv6 addresses in file-backed table(5)s
The file parser splits the line on the ':' character too for key-value
tables, and so mis-parses IPv6 addresses.  The "::1 localhost" example
in table(5) is actually parsed as key "" and value ":1 localhost".

For list tables, the "# @list" marker can be used as a workaround, but
for key-valued the parser has to be fixed.

There are also some weird edge cases when splitting the lines.

Now the parser always splits on the first whitespace or colon, and then
strips the spaces.  For lines starting with '[' the parser will jump to
the matching ']' before attempting to split.  So, for example:

	[::1]:localhost		becomes	"[::1]" -> "localhost"
	[::1] example.org	becomes	"[::1]" -> "example.org"
	foo: bar		becomes "foo" -> "bar"
	foo::bar		becomes "foo" -> ":bar"
	foo : bar		becomes "foo" -> ": bar"

etc...

This only affects the parser for file table(5)s and makemap(8).  Inline
tables or "proc" tables are unaffected.

ok gilles@