Preferred and valid lifetime options are sometimes ignored.
When a prefix is discovered on a network interface and and the IP
address has a valid or preferred lifetime configured that value is
used instead of the static value from the configuration.

Limitation pointed out by & man page text proposed by Ryan Vogt.
Slightly tweaked by me.

rad(8) should calculate the minimum of the static value from the
config file and what is configured on the interface. Implementing that
is slightly complicated and is left for a future diff.