httpd: render "emojis" as text in directory listing
The "up and down" character used in the directory listing is
technically an emoji and can, depending on the browser configuration,
be rendered in colors and whatnot.  set `font-variant-emoji' to ask
the browser to render it as text.  This property, as of now, needs
to be enabled manually in some browsers, but since it's part of the
"CSS Fonts Module Level 4" spec, support for it is expected to be
implemented and turned on "soon" in major browsers.

While here, remove the italic for directories.

Diff from Paul W. Rankin (hello (at) paulwrankin dot com)
ok claudio@