Even though US-ASCII (= ANSI X3.4-1986) only defines 128 characters, the POSIX standard explicitly requires in section 6.2 that "the POSIX locale shall contain 256 single-byte characters", see: https://pubs.opengroup.org/onlinepubs/9799919799/basedefs/V1_chap06.html#tag_06_02
So the current behaviour of treating non-ASCII bytes in an LC_CTYPE=POSIX
input stream as if they were characters is not a POSIX violation, but
actually required by the standard - and not just for awk(1), but for
utility programs in general and even for library functions in general.
Consequently, delete the wrong sentence i added to the STANDARDS section
last year.

Thanks to millert@ and jmc@ for making me realize my mistake.
OK millert@ jmc@