Introduce qwz(4), a work-in-progress port of the Linux ath12k driver.
This driver is not working yet, it's a copy of qwx(4) which bit-by-bit will be
adjusted to work on the newer generation of chips.  Even though this is only a
minor bump over the previous generation, the changes are large enough that
bringup, debugging and long-term maintenance would suffer trying to squash them
into a single driver.  This can be reconsidered once we have reached a stable
state.