Checkpoint of work-in-progress attempt to support the FireGL-UX board. This particular sti(4) device relies upon undocumented callbacks from the kernel, and will default to iodc code otherwise.
The device now initializes correctly, but writes to the frame buffer memory
freeze the system due to bad interaction with the ROM even though it is
disabled at the PCI level - unfortunately I'm currently out of hair to pull
at the moment to make any progress on this.

No functional change intended for other sti(4) devices.