Ensure that file names passed back by readdir do not include a '/' character. The '/' char is the path separator and is not allowed in any filename.
NFS specific report by Apple Security Engineering and Architecture (SEAR).

Input from guenther@ and millert@
from claudio@; OK beck@ miod@

msdos already transfroms for Windows long names a '/' char into '?'.
Do the same for the 8.3 case as well.

This is not ideal since now it is possible that two files in the same
directory have the same name but the msdos code already does a lot of
this and so the problem already exists.

from claudio@; OK beck@ miod@

Do a basic sanity check that dirents returned via fuse are kind of sane.

Ensure that file names passed back by readdir do not include a '/'
character. The '/' char is the path separator and is not allowed in
any filename. On top of this also check that d_reclen and d_namlen
are kind of sane and zero out the padding bytes after d_name.

from claudio@; OK beck@

this is errata/7.4/022_readdir.patch.sig