We aren't ready to choose S0-over-S3 based upon the S0ix bit in FADT. Some machines which work great in S3, don't work great in S0.
Some people want to be able to force S0, mostly for testing purposes
(or to notice improvements as changes are made in the tree).  Provide
a TEMPORARY method via machdep.lidaction=-1 which will be S0-suspend
while =1 remains S3 suspend.  This button will not remain long-term,
but for now, and during 7.6, it will be better than nothing.
ok ratchov kettenis