virtio_pci: Negotiate ACCESS_PLATTFORM feature
Accepting VIRTIO_F_ACCESS_PLATFORM is required for SEV on KVM/qemu.

We always use the standard pci access mechanisms of the
architecture/platform we are running on. Therefore we should negotiate
the VIRTIO_F_ACCESS_PLATFORM feature if it is offered.  Strictly
speaking we should bypass any existing IOMMUs if the host does not offer
this feature, but this no regression and can be fixed later.

feedback from kettenis@