Move psp(4) initialization from attach to device open.
In preparation for automatic loading of the AMD SEV firmware, factor
out the code that initializes the PSP.  As the firmware must be
loaded before the initialization, delay initialization to first use
of the PSP.  Initialization is triggered by opening /dev/psp for
the first time.  A later diff will add firmware loading in a similar
way.
Also provide a ioctl(2) for initialization.  Not used right now.
Will be later used by vmd(8) to reset the PSP.

from hshoexer@; OK mlarkin@