Clean up EC_GROUP_copy()
Switch from artistic free reinterpretations of public API in the same file
to calling the real thing if possible.

This means that we need to copy the group's coefficients first instead of
last, so that we can call EC_GROUP_set_generator() to set - yes - all three
of generator, order, and cofactor of the group.

However, we may not have a generator yet since for some reason it is an
optional field and some code relies on that. In that case simply copy over
order and cofactor and punt on sanity checking for now (since this API
never did that anyway).

Finally set the seed using EC_GROUP_set_seed() instead of using a custom
reimplementation.

ok jsing