Move syslogd(8) TLS handshake callback from read to write handler.
syslogd should immediately detect that the incoming TLS handshake
is complete.  The old logic detected it when the first log message
over TLS was arriving.  For now only a debug message is logged, but
the callback will be used to print the common name of the client
certificate in the future.

OK tb@