bump the type used to specify traffic queue bandwidth to 64bit.
this should let people specify interface and queue bandwidths greater
than ~4Gbit.

this changes the pf ioctls used to specify queues, so if you want
to try this you'll need a new kernel, new headers, and a new pfctl
(and systat). or upgrade using a snapshot. the effort and benefit
of providing compat isn't worth it.

putting it in now so people can kick it around.