Improve error handling of psp(4) driver.
The driver for the AMD platform security processor did use a bunch
of EINVAL error codes.  It is better to propagate the error up from
the subsystem or use more specific values.  Rename the variable
"ret" to "error" to make clear where we deal with an actual errno(2).
Document where and why the error code is ignored.

OK hshoexer@