unwind: support wildcard in blacklist
Any domain in the blacklist that starts with '.', which is not a legal
name due to an empty label, is treated as any subdomain on that zone.

This means that .example.com blocks all requests to any subdomain of
example.com, but allows example.com.

No objections: florian@ OK: kn@