let bpf pick the first attached dlt when attaching to an interface.
this is instead of picking the lowest numbered dlt, which was done
to make bpf more predictable with interfaces that attached multiple
DLTs. i think the real problem was that bpf would keep the list in
the reverse order of attachment and would prefer the last dlt.

interfaces that attach multiple DLTs attach ethernet first, which
is what you want the majority of the time anyway. but letting bpf
pick the first one means drivers can control which dlt they want
to default to, regardless of the numeric id behind a dlt.

ok claudio@