Revert the new rwlock implementation for now.
vfs_busy() uses RW_SLEEPFAIL in a broken way. It is possible that the
object holding the rwlock is freed while other processes are sleeping
on this lock. This worked before by luck and no longer does now since
part of the struct needs to be updated after the sleep.
vfs_busy() needs to be fixed but that will take a bit of time.
OK dlg@