Implement two-level (indirect) Device Table support to increase the range of DeviceIDs we can set up for translation.
Peripherals capable of doing DMA/MSIs are supposed to show up with
unique DeviceIDs.  The Device Table maps the DeviceID to a Interrupt
Translation Table.  So far we only used a single contiguous block for
the Device Table, but on some machines this does not cover the whole
range of physical devices.  Using the GIC's indirect mode allows to
move to a two-level setup to increase the range.

ok kettenis@