In prefix_adjout_update() and prefix_adjout_withdraw() only link prefixes to the update / withdraw queue if the peer is up.
For updates the Adj-RIB-Out is already updated and nothing more needs to happen.
For withdraws the prefix is unlinked and needs to be destroyed if the peer is
not up. For this to work correctly make the prefix as dead.

Right now this is a no-op since the functions are only called if the peer
is up.
OK tb@