Improve cleanup process
In some cases, following a fallback from RRDP to RSYNC, files are
moved to the wrong place. With this change, only rsync repositories
are considered when looking up where to move DIR_TEMP files.

repo.c requires more study, this fix might be a bandaid for a problem
not fully understood.

"put it in" claudio@ OK tb@