Add an .St argument for C23.
The mismatch of the year numbers ist not a typo.
The official name is "ISO/IEC 9899:2024", so the "4" is correct there.

But as a colloquial name, "C23" is more widespread than "C24",
probably because __STDC_VERSION__ == 202311L: The final version
of the standard was ready in 2023, only formally publishing it
took about 11 months after it was ready.

The main reason for choosing "-isoC-2023" as the .St argument
is to match what NetBSD already committed to their tree.
Also, it aligns well with the existing "-isoC-2011".

Based on an incorrect patch from Thomas Klausner <wiz at netbsd dot org>,
fixed by me.

OK jmc@