Run TCP SYN cache timer with shared net lock.
Reset the inp_socket pointer in in_pcbdetach() when a socket is
about to be freed.  By protecting this with a mutex, it is possible
to reliable retrieve the socket from an inpcb reference.  By using
the socket refcount of the listen socket, switch from mutex to
socket lock in syn_cache_timer().

tested by Hrvoje Popovski; OK mvs@