Reference count the inpcb in TCP timers.
Switch from struct tcpcb to inpcb in the TCP timer argument.  The
latter already has a reference counter.  Increment it at timeout_add()
and decrement at timeout_del() or when handler runs.  The reaper
timeout is special as it does not need a reference, the inpcb is
already dead.  Use special field t_timer_reaper instead of regular
TCP timeout and run it without reference or lock.

OK mvs@