Do not depend on RW_SLEEPFAIL in ntfs_ntlookup
While it is a cute optimisation it makes no sense to be a special
snowflake since there is no parallelism in this code. Instead use
the same pattern used by the other filesystems that detects a
collision in ntfs_nthashins() followed by a retry of lookup.

Tested by ian@ and jmatthew@
OK mpi@