let pppoe data packets go through if_vinput instead of the pppoeinq.
provide pppoe_vinput() for ether input to call. if the packet is
for data in an established pppoe session, it can push it straight
into the stack with if_vinput. otherwise pppoe_vinput returns the
packet to ether_input, which can queue it for processing with the
existing code.

this should improve throughput and reduce jitter for pppoe input,
and there's some evidence that it reduces packet loss.

tested by maurice janssen and myself.