Finally include ARIN's RPKI Trust Anchor Locator
Backstory:

We strive to provide a source distribution that can be freely
used, copied, modified, and redistributed by anyone and for any
purpose. Unfortunately, up until now, several legal barriers
stopped us from distributing the ARIN TAL.

The main legal obstacles stemmed from the terms and conditions
in the Relying Party Agreement (RPA) governing access to ARIN's
RPKI repository and the way ARIN tried to ensure that agreement
was binding.

Originally, ARIN used a cumbersome email-based method of RPA
acceptance. Then in 2016, ARIN moved to a browser user
interface-based "clickwrap" method via their website, requiring
affirmative assent via a mouse-click. In 2019, a new approach
was offered where ARIN encouraged software distributors to pass
through restrictive terms in irksome pop-ups during installation;
this we also deemed unworkable. In 2022, this "Redistributor
RPA" was subsumed in a single RPA, but without material change
to the mechanics, and the terms remained a showstopper.

Over the years various people suggested ARIN to adopt an "AS IS"
disclaimer of warranties. In 2024, Job Snijders authored a
modified version of the BSD license and proposed ARIN to include
it in the optional comment section in the TAL file format.
In 2025 ARIN updated its TAL to include this disclaimer.

After more than a decade of iterations, this public key now is
available to the public in an unencumbered fashion and can serve
its main purpose, a purpose that can only occur through
immensely widespread distribution.

Some pointers for historic context:

2012 - https://web.archive.org/web/20130127143807/https://www.arin.net/resources/rpki/rpa.pdf
2015 - https://web.archive.org/web/20150203184532/https://www.arin.net/public/rpki/tal/index.xhtml
2017 - https://lists.arin.net/pipermail/arin-ppml/2017-January/031231.html
2017 - https://www.arin.net/vault/about/welcome/board/meetings/20170405/exhibit_d.pdf
2018 - https://mailman.nanog.org/pipermail/nanog/2018-September/097161.html
2019 - https://papers.ssrn.com/sol3/Papers.cfm?abstract_id=3308619
2019 - https://www.arin.net/vault/participate/meetings/reports/ARIN_43/PDF/PPM/yoo_rpki.pdf
2019 - https://www.arin.net/vault/announcements/20191021/
2019 - https://www.theregister.com/2019/10/28/arin_rpki_open_source/
2022 - https://www.arin.net/vault/announcements/20220926/
2022 - https://www.arin.net/vault/announcements/20220929/
2025 - https://www.arin.net/announcements/20250116-tal/

OK tb@ claudio@ deraadt@