Do not unlock the socket in soabort().
One difference between UNIX and internet sockets is that UNIX sockets
unlock in soabort() while TCP does not do that.  in_pcbdetach()
keeps the lock, change uipc_abort() to behave similar.  This also
gives symetric lock and unlock in the caller.  Refcount is needed
to call unlock on an aborted socket.
The queue 0 in soclose() is only used by UNIX sockets, so remove
the "if" persocket.  The "kassert" persocket in soisconnected() is
not needed.

OK mvs@