add options to interactive sdiff(1) merge for choosing both sides of a diff
In some situations there is no clear winner of one side of the diff over
the other, and both sides are needed in the merge result.
Add new merge commands which allow both sides to be chosen in either order,
without a requirement to run the lines through an editor via the existing
'e b' merge command.

ok sthen@