Split union RT_gw into separate struct rtentry fields.
The union of rt_gwroute and rt_cachecnt makes debugging harder.  In
theory the RTF_GATEWAY flag distinguishes which field is relevant,
but that does not help when hunting bugs and races in this area.
The benefit of saving 4 bytes in struct rtentry is minimal.
Break up the RT_gw union.  To keep memory compact, sort pointers
to the beginning and integres to the end of struct rtentry.

OK mvs@