In VI command line editing mode, make the behaviour of "1P" consistent with the behaviour of "2P" such that both back up to the first byte of the last character inserted.  As a side benefit, that also makes "[n]P" work in the presence of UTF-8 characters.
According to pascal@, the new behaviour of "P" is conistent with ksh93
and all historical versions of ksh he tested on sdf.org, so the deviant
behaviour of "1P" was likely not intentional, but a bug in pdksh.
According to Walter Alejandro Iglesias <wai at roquesor dot com>,
the new behaviour is also conistent with vim and bash.

OK pascal@