Be explicit about rdns_count.
This is equivalent to the check in the previous lines, but it is hard
to reason about because it depends on struct sockaddr_rtdns sized just
correctly.

The previous check has to stay, because it makes sure we have enough
space to send our nameservers to the kernel in case struct
sockaddr_rtdns ever changes size.

This tripped up claudio during a review.