allocate pf states and state keys on cache lines.
the idea is to try and avoid false aliasing of pf states as we
become more parallel in the future. i also want this as a reminder
to stop wasting my time optimising for memory usage.

i've been running this for years without issue.

ok sashan@ (in mar 2023). he agrees that if it sucks we can pull it out.