pkgconf: initial pledge from upstream
This is a very early pledge "stdio rpath wpath cpath" since the utility
has to walk the fs to read .pc files, on --silence-errors it redirects
stderr to /dev/null, and it can be asked to create a log file.

Once refactored to accomodate unveil, this might be tightened further.
For now this is much better than nothing. Went through an amd64 bulk.

ok schwarze semarie