In ksh(1) VI mode, prevent display corruption when the command line being edited starts with a UTF-8 continuation byte.
In particular, in case a UTF-8 continuation byte is the first byte
on the line, count it as width 1 rather than width 0, and in case
a UTF-8 continuation byte is the first byte that changes, only
back up the column if it's not in the first column.
This was discussed in more detail on tech@.

OK lucas@;
anton@ and kn@ also think i should go ahead.