Don't use the array index as the group_id
This is a precursor to adding new group ids for post quantum
stuff which are up in the 4000 range, so using the array index
as the group id will be silly.  Instead we just add the group
id to the structure and we walk the list to find it.

This should never be a very large list for us, so no need
to do anything cuter than linear search for now.

ok jsing@, joshua@