Change the way the update interval is set
Instead of setting the value through a 'prefix' to the interactive
command, make this like the 's' command in top(1). Set the default
interval to 1 second.

While there, get rid of the '[prefix]g' movement: 'g' now just jumps
to the top; also don't print the current position.

OK kn@