Handle IOIO for AMD SEV-ES VMGEXIT in vmm(4).
Fill in the actual handler for IOIO related GEXIT.  The actual "in"
or "out" instruction is encoded in the exitinfo provided by the
guest in the GHCB.  For "in" vmm(4) returns the result in the A
general purpose register.

The actual emulation is done by svm_handle_inout() and vmd(8).  On
the way back into the guest, vmm(4) synchronizes the computed result
back to the GHCB.

With this commit, vmm(4) and vmd(8) will be able to deal with all
GEXITs that will be generated for now by openbsd SEV-ES enabled
guest.

from hshoexer@; OK mlarkin@