Handle AMD SEV-ES GHCB MSR protocol for CPUID in vmm(4).
When a SEV-ES guest is in locore0, it will not be able yet to use
a GHCB to communicate with vmm(4).  Therefore, AMD specifies a "GHCB
MSR protocol" that uses the lower 12 bits of the GHCB MSR to request
services from vmm(4).  Guest writes to the GHCB MSR will show up
in the v_ghcb_gpa member of the VMCB and are thus accesible by
vmm(4).  The response of vmm(4) can be provided by writing it to
the VMCB.

In locore0 a SEV-ES guest will need to request CPUID from vmm(4)
using the GHCB MSR protocol.  This commit provides vmm(4) implementation.

from hshoexer@; OK mlarkin@