Collect uvm_swapout_threads(), pmap_collect() and __HAVE_PMAP_COLLECT and take them behind the shed.
The uvm code used to call into pmap_collect is overly optimistic and
thinks that a process that has all its thread sleeping will remain like
that for while without holding any relevant locks.

This is combined with the fact that pmap_collect functions are not MP safe.
The result are bad use-after-free errors. Since only a few archs implement
this feature it is probably better to just remove it then trying to fix
all of the issues.

OK mpi@, kettenis@, deraadt@