.Lb libalpha ; OK matthieu@ tedu@.
There is no .In macro because application code using these functions needs
to copy the required prototypes from the manual page into the code instead
of using any #include directive.

According to matthieu@, this "was used by the X server on alpha, which
has been broken for several years.  No one seems to care enough to try
to figure out what is needed to fix it."