Unlock IPCTL_MTUDISC case of ip_sysctl().
`ip_mtudisc' is atomically accessed boolean, so allow only 0 and 1
values to set. Also, while `ip_mtudisc' is 0, the rt_timer_queue_flush()
will be triggered all the times even if we do read access. There is no
reason for that, so flush the queue only if this thread successfully
assigned 0 value. The rt_timer_queue_flush() requires to be serialized
with the netlock.

ok bluhm