rpki-client: add an AIA handler fit for all certs
cert_aia() is similar to cert_sia(), but it is easier. It does not
the slight inaccuracy of x509_get_sia() and allows multiple caIssuers
access methods. Like for the SIA case, we remember the first rsync
access method and ignore all other caIssuers. Also like in the SIA
case, we'll dedup the code a bit laster.

ok job