hidmt: separate contact inputs correctly
When processing input reports, hidmt_input identifies groups of contact-
level items by checking for the repetition of an item type - the first
type it has encountered in the report.  This fails for reports starting
with report-level items, which only occur once.  The 'firstu' value must
not be set to a report-level type.

Thanks to Thomas Dettbarn for analyzing and reporting the problem.

ok stsp@