rpki-client: print zero signtime
This is a leftover from the time where we had to allow signtime to be
absent. cms_parse_validate_internal() has rejected signed objects
without signing-time since cms.c r1.43 (April 2024). The signed object
parsers all set it. If signtime happens to be 0, it is because the
signed object is weird. No reason to treat such an object specially at
this point in the code.

ok job