revert 'fix wrong ifmedia macro used to check for IFM_AUTO'; change was wrong
While IFM_AUTO is indeed in the IFM_SUBTYPE range of the ifmedia word,
what this check is really trying to find out is whether any bits in
the IFM_MODE range are set. Those bits indicate whether userland has fixed
the phy mode to 11a/11b/11g etc. And because IFM_AUTO is zero this check
works as intended, even though comparing IFM_MODE range values to a value
declared in the IFM_SUBTYPE range is questionable coding style.