Remove lapic TPR remnants on amd64
On amd64, the TPR register is also available via cr8. We use lcr8(0) to
set it to 0 in the appropriate places. There is no need to touch it
in lapic_map().

This means the only remaining symbol from the lapic MMIO page we use is
local_apic. All other registers are expressed in offsets from that symbol via
i82489_writereg/i82489_readreg(). Remove the other symbols.

ok guenther@