Pass down existing ifp pointer to nd6_free() and rtrequest_delete().
Coverity complained about useless ifp NULL pointer check in
rtrequest_delete() and a later unconditional dereference.  For every
caller except nd6_free() it was clear that ifp was a valid interface.
Pass down ifp to nd6_free() instead of calling if_get().  Then it
is obvious that ifp is not NULL and can be used in all callees.

CID 1648414
input and OK mvs@