Remove null_prologue() and the definition a "return" insn. That particular insn is used to be able to perform some optimizations by avoiding to jump to the function's epilogue, if a short return form is possible.
Theoretically, this would be the case on m88k for leaf functions with no
locals, which would not have adjusted the stack pointer.

But due to stack alignment constraints and the way the MI code sets up
local variables, this turns out never to be the case, so optimizations
based on this insn can never suceed. By removing the insn declaration the
code attempting to perform these optimizations will no longer get
compiled in.