vmd(8): load the full SeaBIOS image in lower bios memory.
My change to make space for edk2 firmware images trimmed the amount
of the image we load into the lower bios area. Instead of trimming
to 128KB, trim to the bios area size. In the case of SeaBIOS this
loads the full image.

This fixes booting certain Linux guests that use boot loaders that
(I believe) try calling into parts of the bios in this lower memory.
(We don't shadow upper and lower bios regions.) It puts them typically
into a spiral of hammering the PIC.