Improve externs for KEYMAPE()
Gentoo tried to compile mg with -flto, and in doing so they discovered
that some extern in .c files resulted in the use of a wrongly sized
anonymous struct. Nowadays the KEYMAP and KEYMAPE() kludge should probably
be replaced with a flexible array member in a single type, but that's a
slightly bigger change that requires more thinking than this diff.

From the Gentoo bug tracker via Han Boetes