correct family test when setting Zenbleed chicken bit on i386
For AMD Zen 2 based systems a microcode update is the preferred way of
mitigating Zenbleed.  An MSR to set a chicken bit is another way.  The
CPU family test around this MSR was wrong.

When testing for CPU family 0x17 the extended value must be checked.
A non-extended family id of 0x17 is impossible as the mask is & 15.
So the previous test would never match.

In i386 identifycpu(), the function local family variable is always the
non-extended value.  This problem does not apply to amd64.  The amd64
change to move this part of identifycpu() to cpu_fix_msrs() uses an
extended family id.

ok bluhm@ deraadt@ mlarkin@