implement draft-ietf-acme-profiles for acme-client, ok florian
letsencrypt uses this to allow asking for a certificate with a specific
TLS profile; see https://letsencrypt.org/docs/profiles/ for current options

note that with current timers, if you select the non-default "shortlived"
profile, renewal will be done at each acme-client run. if this results in
exceeding rate limits, adjust cronjobs, or don't do that. (this is being
looked at, but may possibly be too sensitive to change before release).