Adapt renewal calculation for shortlived certificates.
If the lifetime is more than 10 days renew if less than 1/3 of the
lifetime is left. Otherwise renew after 1/2 of the remaining lifetime.

Since we suggest to run the cronjob daily, this is capped at 3 days
remaining lifetime to have the opportunity to run the cronjob at least
twice.

Input & OK tb, sthen

Putting it in now because it gives buypass users 60 days of warnings
instead of 30 that their certificate can't be renewed (pointed out by
sthen).