Lower bound for how long to prefer IPv6 at least.
RFC 8925 suggests 5 minutes, so that's what we are going with.

kn noticed that if the IPv6-Only preferred option is send from the
server with a value of 0, dhcpleased(8) would still request a lease
while apple devices would not.

There were two problems with that:
1. A value of 0 would be treated as turning the feature of, which is
not correct.
2. The value would be used for a timeout, so a very small value would
create a timeout that would constantly fire.

OK kn