with DLT_ENC, use the first nibble of the ip to pick the ip version.
this is because the enc link type header which is supposed to have
an address family field is populated by the kernel in the hosts
native endian when it should be in a consistent order no matter
what type of cpu it is run on.

this re-uses the DLT_RAW handling to look at the ip version field.