extend and add structs for ioctls to manage a vlan aware bridge.
this sneaks a pvid (port vlan id, or the native vlan id) into padding
in struct ifbreq.

make a new version of struct ifbareq called ifbvareq that can be
used to add an address to the bridge forwarding database with a vid
(vlan id).

add a struct ifbrvidmap to manage the allowed vlan tags on a port
in a bridge.

these will be used by veb(4).