fix a panic when autodial (link1) on pppoe tries to run.
"dialling" needs the kernel lock because thats what protects the
ppp state machine transitions. sppp_output tries to bring the ppp
link up if it's asked to send a packet, but it can be called without
the kernel lock now.

sppp_output uses a task in the systq to bring the link up now.

reported by illya.meyer@wiesan.de on bugs@