Duplicate struct prefix into struct prefix_adjout and adjust code to work with that.
Make sure that anything in the adj-rib-out uses the new struct. In
some cases code has to be duplicated since it is no longer possible to
use one code path to cover both Local-RIB and Adj-Out-RIB. This is
mostly in the code handling the 'bgpctl show rib' family of commands.

OK tb@