Rework the interface to rde_generate_updates() and up_genrate_XYZ() to pass the old prefix as just the path_id_tx identifier.
Only up_generate_addpath_all() actually uses this information and there
this is enough to find the affected prefix in the adj-rib-out.
Also adjust the order of operation in up_generate_addpath_all() so the
prefix is not first removed and readded for the case where a prefix
is simply updated.

Rework the code in prefix_evaluate_nexthop() to be much more like
prefix_evaluate(), it should be possible to factor out common code at
some point.  In peer_add() ensure that path_id_tx can't be 0. Since 0
is now reserved for no-such-path.

OK tb@