Fix vmd reboots broken by my pause barrier changes.
Reported by bluhm@.

Long story is vmd has some complexity in how the decision to "reboot"
a vm process is made. The pause barrier diff clobbered a return
value that bubbled up to a value passed to exit(3) and picked up
by the parent vmd process (the "vmm" process) that determines if a
process exit should result in relaunching the vm process or not.

For now, just stop clobbering the value and don't report the error
returned by pthread_barrier_destroy(3).