fix srp_follow to close a window on use-after-free
Use srp_enter() to get a new reference to the next element while
keeping the current element alive. Afterwards the old reference can
safely be released and the hazard in the caller provided srp_ref
struct can be updated to the hazard of the new element.

This is just in time for almost all the SRP code in the tree to go away.

from Carsten Beckmann carsten_beckmann at genua.de
ok jmatthew@