Rearrange command completion so callbacks are called without holding any locks. This makes it possible to mark the interrupt handler MPSAFE, but we're not actually doing that yet.
Releasing the cq mutex means the completion callback can't use the cq
entry, so we have to copy any fields we use from it into the ccb. For now,
that's just the flags. This simplifies the callbacks in a few places.

ok dlg@ (some time ago)
also tested by kettenis@ with aplns(4)