Ampere Altra ACPI IORT table does not define SPIs for the different IRQ types.  Instead I assume we'll have to make use of MSIs, which I'll yet have to implement.
It's not much of a problem though because these interrupts are mostly used
in case something goes wrong, e.g. due to an invalid access; if everything
is working fine, we don't need interrupts right now.

Unfortunately though we bailed out at an unfortunate moment:  After we had
already initialized the SMMU to intercept, but before the driver is hooked
up as IOMMU, thus not mapping anything but the SMMU enforcing boundaries.

For now we can simply log that the interrupts are not there and proceed.
Note that SMMU on ACPI isn't enabled by default yet anyway.